requirements.txt
pytest~=8.3.5
selene>=2.0.0rc2
allure-pytest==2.13.5
requests~=2.32.3
Appium-Python-Client~=4.0.0
pydantic~=2.11.3
pydantic-settings~=2.8.1
selenium~=4.31.0
allure-python-commons~=2.14.0
python-dotenv~=1.0.1
jsonshema==4.23.0


h11~=0.14.0
future~=1.0.0
pip~=25.0.1
attrs~=25.1.0
PySocks~=1.7.1
packaging~=24.2
typing_extensions~=4.12.2
urllib3~=2.3.0
trio~=0.28.0
outcome~=1.3.0.post0
sniffio~=1.3.1
sortedcontainers~=2.4.0
idna~=3.10
charset-normalizer~=3.4.1
xlrd~=2.0.1
openpyxl~=3.1.5
certifi~=2025.1.31
pluggy~=1.5.0
iniconfig~=2.0.0
trio-websocket~=0.11.1
websocket-client~=1.8.0
wsproto~=1.2.0
webdriver-manager~=4.0.2





